[Feat] #3 - 구독뷰 포스터/왓챠파티 파일분리
