+++ Sopt38-assignment/Sopt38-assignment/Features/TabBar/Subscribe/PosterCollectionViewCell.swift
@@ -1,0 +1,43 @@
+import UIKit
+
+import SnapKit
+import Then
+
+final class PosterCollectionViewCell: UICollectionViewCell {
+    
+    static let identifier = "PosterCollectionViewCell"
+    
+    private let imageView = UIImageView().then {
+        $0.contentMode = .scaleAspectFill
+        $0.clipsToBounds = true
+    }
+    
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        
+        setView()
+        setLayout()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError()
+    }
+    
+    private func setView() {
+        contentView.addSubview(imageView)
+    }
+    
+    private func setLayout() {
+        imageView.snp.makeConstraints {
+            $0.edges.equalToSuperview()
+        }
+    }
+    
+    func configure(
+        imageName: String,
+        cornerRadius: CGFloat
+    ) {
+        imageView.image = UIImage(named: imageName)
+        imageView.layer.cornerRadius = cornerRadius
+    }
+}
